BRK KOT class for shared CSV formatting methods.
